vtworkerclient: Move client test suite into its own package.
Before this change production binaries, which included vtworkerclient, unnecessarily also included all dependencies (e.g. the go/zk package) of the test suite.

vtctlclient has a similar structure. There we had already a separate package only for the client test suite. I've added comments in its package as well to document why we have a separate package.
